drop IrcLegacySubscriber
you should migrate logic contained to your workflow methods
added to examples
