+++ src/Extension/IrcNotifyExtension.php
@@ -13,7 +13,6 @@
 
 namespace Eventum\Extension;
 
-use Eventum\Event\IrcLegacySubscriber;
 use Eventum\Event\IrcSubscriber;
 use Setup;
 
@@ -29,8 +28,6 @@ class IrcNotifyExtension extends AbstractExtension
         $setup = Setup::get();
         if ($setup['irc_notification'] === 'enabled') {
             $subscribers[] = IrcSubscriber::class;
-            // @deprecated IrcLegacySubscriber will be removed in 3.5.0 along with irc bot code
-            $subscribers[] = IrcLegacySubscriber::class;
         }
 
         return $subscribers;
